refactor(member) : 패키지 분리 및 위치 변경
