+++ src/main/java/roomescape/member/dao/MemberDao.java
@@ -4,7 +4,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.member.Member;
+import roomescape.member.domain.Member;
 
 @Repository
 public class MemberDao {
