+++ src/main/java/roomescape/member/dto/response/MemberResponse.java
@@ -1,4 +1,4 @@
-package roomescape.member;
+package roomescape.member.dto.response;
 
 public class MemberResponse {
     private Long id;
