+++ src/test/java/roomescape/MemberDaoTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
-import roomescape.member.Member;
+import roomescape.member.domain.Member;
 import roomescape.member.dao.MemberDao;
