+++ src/test/java/roomescape/MissionStepTest.java
@@ -31,7 +31,6 @@ public class MissionStepTest {
         RestAssured.port = port;
     }
 
-
     @Test
     void 일단계() {
         Map<String, String> params = new HashMap<>();
